added more cleanups
